Split out types and components
